ensure cell ids are saved as int values
otherwise they can not be matched to their corresponding locations in the segmentation masks

+++ src/scportrait/pipeline/classification.py
@@ -552,7 +552,9 @@ class _ClassificationBase(ProcessingStep):
             # save nucleus segmentation
             obs = pd.DataFrame()
             obs.index = obs_indices
-            obs["instance_id"] = obs_indices
+            obs["instance_id"] = obs_indices.astype(
+                int
+            )  # this needs to be int otherwise data will not be loaded correctly
             obs["region"] = f"{mask_type}_{self.MASK_NAMES[0]}"
             obs["region"] = obs["region"].astype("category")
 
@@ -574,7 +576,9 @@ class _ClassificationBase(ProcessingStep):
             # save cytoplasm segmentation
             obs = pd.DataFrame()
             obs.index = obs_indices
-            obs["instance_id"] = obs_indices
+            obs["instance_id"] = obs_indices.astype(
+                int
+            )  # this needs to be int otherwise data will not be loaded correctly
             obs["region"] = f"{mask_type}_{self.MASK_NAMES[1]}"
             obs["region"] = obs["region"].astype("category")
 
@@ -1200,7 +1204,7 @@ class _cellFeaturizerBase(_ClassificationBase):
 
             obs = pd.DataFrame()
             obs.index = obs_indices
-            obs["instance_id"] = obs_indices
+            obs["instance_id"] = obs_indices.astype(int)
             obs["region"] = f"{mask_type}_{self.MASK_NAMES[0]}"
             obs["region"] = obs["region"].astype("category")
 
@@ -1236,7 +1240,7 @@ class _cellFeaturizerBase(_ClassificationBase):
 
             obs = pd.DataFrame()
             obs.index = obs_indices
-            obs["instance_id"] = obs_indices
+            obs["instance_id"] = obs_indices.astype(int)
             obs["region"] = f"{mask_type}_{self.MASK_NAMES[1]}"
             obs["region"] = obs["region"].astype("category")
